buildtools/burg/plank.c(109): remark #2259: non-pointer conversion from "int" to "short" may lose significant bits
http://monetdb.cwi.nl/testing/projects/monetdb/testweb/logs/39245:124139f4ff4a/Int-Fedora-i386-propcheck/make.html#l481
http://monetdb.cwi.nl/testing/projects/monetdb/testweb/logs/39245:124139f4ff4a/Int-Fedora-i386-propcheck/make.html#l485

+++ buildtools/burg/plank.c
@@ -106,8 +106,8 @@ newException(index, value) int index; int value;
 	Exception e;
 
 	e = (Exception) zalloc(sizeof(struct except));
-	e->index = index;
-	e->value = value;
+	e->index = (short)index;
+	e->value = (short)value;
 	return e;
 }
